+++ src/plugins/alanya-live-bot/server/utils/components/keyboardCommands/search/repeatSearchCars.js
@@ -1,4 +1,4 @@
-const { userLang } = require('../../../../../../botUtils/botsLanguages');
+const { userLang } = require('../../../../../../botUtils/language');
 const getUser = require('../../../../../../botUtils/userController');
 const searchCarsFunc = require('./searchCars');
